requirements.txt
asgiref==3.7.2
boto3==1.28.63
botocore==1.31.63
Django==4.2.6
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==2.0.1
